更新

+++ utils/rresult/returnResult.go
@@ -1,7 +1,8 @@
-package rresult 
+package rresult
+
 type Result struct {
-	Code int `json:"code"`
-	Message string `json:"message"`
-	Data interface{} `json:"data"`
-	Totoal int `json:"totoal"`
+	Code    int         `json:"code"`
+	Message string      `json:"message"`
+	Data    interface{} `json:"data"`
+	Totoal  int64       `json:"totoal"`
 }
